script_random_forest.py
One of the models I built to predict the price. For this script, I only select all continuous variables and a few categorical variables as the predictors (so not all features included in this script), I will try to include more categorical features in the next step of the project. I use Random Forest algorithm in this script, with parameter tuning is done using GridSearchCV by Python Scikit-Learn.